+++ rebuild/src/public_key.h
@@ -1,0 +1,27 @@
+#ifndef BLST_TS_PUBLIC_KEY_H__
+#define BLST_TS_PUBLIC_KEY_H__
+
+#include <memory>
+
+#include "addon.h"
+#include "blst.hpp"
+#include "napi.h"
+
+#define BLST_TS_PUBLIC_KEY_LENGTH_COMPRESSED 48U
+#define BLST_TS_PUBLIC_KEY_LENGTH_UNCOMPRESSED 96U
+
+class PublicKey : public Napi::ObjectWrap<PublicKey> {
+   public:
+    bool _has_jacobian;
+    bool _has_affine;
+    std::shared_ptr<blst::P1> _jacobian;
+    std::shared_ptr<blst::P1_Affine> _affine;
+
+    static void Init(Napi::Env env, Napi::Object &exports, BlstTsAddon *module);
+    static Napi::Value Deserialize(const Napi::CallbackInfo &info);
+    PublicKey(const Napi::CallbackInfo &info);
+    Napi::Value Serialize(const Napi::CallbackInfo &info);
+    Napi::Value KeyValidate(const Napi::CallbackInfo &info);
+};
+
+#endif /* BLST_TS_PUBLIC_KEY_H__ */
